changed image hosting to dropbox

+++ src/store/request/mutations.ts
@@ -1,11 +1,14 @@
 import { MutationFunction } from "react-query";
 import firebase from "firebase";
 
+import { isNil } from "lodash";
+
 import { RequestItemImage, RequestObject } from "../../types/request-types";
 import { firestore, storageRef } from "../../firebase/localFirebase";
 import { REQUESTS_COLLECTION } from "../../firebase/constants";
 import { handleError } from "../../utils/handleError";
 import { log } from "../../utils/log";
+import { deleteFile, uploadFile } from "../../fileStorage/fileStorage";
 
 const RETRIES_LIMIT = 10;
 
@@ -13,39 +16,43 @@ export const uploadImageMutation: MutationFunction<
   RequestItemImage,
   RequestItemImage
 > = async (image) => {
-  const metadata = {
-    contentType: "image/jpeg",
-  };
-
   try {
-    if (image.file !== undefined) {
-      for (let i = 0; i < RETRIES_LIMIT; i++) {
-        const name = image.file.name ?? image.name;
-        const imageSnap = await storageRef
-          .child(name)
-          .put(image.file as File, metadata);
-
-        log(
-          `image.file.size: ${image.file.size}`,
-          `imageSnap.bytesTransferred: ${imageSnap.bytesTransferred}`,
-          `name: ${name}`,
-          `metadata: ${metadata}`
-        );
-
-        if (
-          imageSnap.state === "success" &&
-          image.file.size === imageSnap.bytesTransferred
-        ) {
-          const url = await imageSnap.ref.getDownloadURL();
-
-          return { url, name };
-        }
-      }
-    }
+    return await uploadFile(image);
   } catch (e) {
     handleError(e);
-    console.log("e", e);
-    throw e;
+    try {
+      const metadata = {
+        contentType: "image/jpeg",
+      };
+      if (image.file !== undefined) {
+        for (let i = 0; i < RETRIES_LIMIT; i++) {
+          const name = image.file.name ?? image.name;
+          const imageSnap = await storageRef
+            .child(name)
+            .put(image.file as File, metadata);
+
+          log(
+            `image.file.size: ${image.file.size}`,
+            `imageSnap.bytesTransferred: ${imageSnap.bytesTransferred}`,
+            `name: ${name}`,
+            `metadata: ${metadata}`
+          );
+
+          if (
+            imageSnap.state === "success" &&
+            image.file.size === imageSnap.bytesTransferred
+          ) {
+            const url = await imageSnap.ref.getDownloadURL();
+
+            return { url, name, uploaded: true };
+          }
+        }
+      }
+    } catch (e) {
+      handleError(e);
+      console.log("e", e);
+      throw e;
+    }
   }
 
   return Promise.reject();
@@ -55,7 +62,14 @@ export const deleteImageMutation: MutationFunction<
   void,
   RequestItemImage
 > = async (image) => {
-  await storageRef.child(image.name).delete();
+  try {
+    if (!isNil(image)) {
+      await Promise.allSettled([
+        storageRef.child(image.name).delete(),
+        deleteFile(image.name),
+      ]);
+    }
+  } catch (e) {}
 };
 
 export const sendRequestMutation: MutationFunction<
